Extract common settings persist logic to the separate react hook
