+++ client/web/src/insights/core/settings-action/edits/index.ts
@@ -1,0 +1,2 @@
+export * from './edits'
+export { applyEditOperations } from './apply-edit-operations'
